requirements.txt
fastapi
uvicorn[standard]
requests
jinja2
pytest
python-dotenv